requirements.txt
amqp==5.1.1
anyio==3.7.0
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.3.0
bleach==5.0.1
blinker==1.4
boto3==1.26.146
botocore==1.29.146
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dbus-python==1.3.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.0
Django==4.2.1
django-apscheduler==0.6.2
django-bleach==3.0.1
django-celery-results==2.5.1
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-environ==0.10.0
django-excel-response2==3.0.5
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
excel-base==1.0.4
exceptiongroup==1.1.1
executing==1.2.0
Faker==18.10.1
fastjsonschema==2.16.3
filelock==3.12.1
fonttools==4.40.0
fqdn==1.5.1
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.20.2
httptools==0.5.0
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
isoweek==1.3.3
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==9.4.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==8.10.0
msgpack==1.0.5
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
oauthlib==3.2.2
packaging==23.0
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipx==1.2.0
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.7.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.6.1
screen==1.0.1
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
systemd-python==234
terminado==0.17.1
TimeConvert==3.0.13
tinycss2==1.1.1
tomli==2.0.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.16
userpath==1.8.0
uvicorn==0.22.0
vine==5.0.0
wadllib==1.3.6
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
wordcloud==1.9.2
xlwt==1.3.0
zipp==1.0.0
